backend/telegram-bot/requirements.txt
aiogram>=3.0
psycopg2-binary
